services/nlp/requirements.txt
spacy>=3.7.0
kafka-python>=2.0.2
